Try to simplify Section and Range. Remove Range.VLEN.
