+++ core/src/main/kotlin/com/sunya/cdm/api/Cdl.kt
@@ -1,6 +1,8 @@
 package com.sunya.cdm.api
 
 import com.sunya.cdm.util.Indent
+import com.sunya.cdm.util.escapeCdl
+import com.sunya.cdm.util.escapeName
 import java.nio.ByteBuffer
 import java.util.*
 
@@ -103,48 +105,4 @@ fun Attribute.cdl(varname: String, indent : Indent = Indent(2)) : String {
 
 internal fun ByteBuffer.toHex() : String {
     return "0X" + HexFormat.of().withUpperCase().formatHex(this.array())
-}
-
-/////////////////////
-private val org = charArrayOf('\b', '\n', '\r', '\t', '\\', '\'', '\"')
-private val replace = arrayOf("\\b", "\\n", "\\r", "\\t", "\\\\", "\\'", "\\\"")
-
-fun escapeName(s: String): String {
-    return replace(s, org, replace).replace(" ", "_")
-}
-
-fun escapeCdl(s: String): String {
-    return replace(s, org, replace)
-}
-
-fun replace(x: String, replaceChar: CharArray, replaceWith: Array<String>): String {
-    // common case no replacement
-    var ok = true
-    for (aReplaceChar in replaceChar) {
-        val pos = x.indexOf(aReplaceChar)
-        ok = pos < 0
-        if (!ok) break
-    }
-    if (ok) return x
-
-    // gotta do it
-    val sb = StringBuilder(x)
-    for (i in replaceChar.indices) {
-        val pos = x.indexOf(replaceChar[i])
-        if (pos >= 0) {
-            replace(sb, replaceChar[i], replaceWith[i])
-        }
-    }
-    return sb.toString()
-}
-
-fun replace(sb: StringBuilder, remove: Char, replaceWith: String) {
-    var i = 0
-    while (i < sb.length) {
-        if (sb[i] == remove) {
-            sb.replace(i, i + 1, replaceWith)
-            i += replaceWith.length - 1
-        }
-        i++
-    }
 }
