+++ core/src/main/kotlin/com/sunya/cdm/util/Escaping.kt
@@ -29,3 +29,47 @@ fun makeValidCdmObjectName(orgName: String): String {
         }
     }
 }
+
+/////////////////////
+private val org = charArrayOf('\b', '\n', '\r', '\t', '\\', '\'', '\"')
+private val replace = arrayOf("\\b", "\\n", "\\r", "\\t", "\\\\", "\\'", "\\\"")
+
+fun escapeName(s: String): String {
+    return replace(s, org, replace).replace(" ", "_")
+}
+
+fun escapeCdl(s: String): String {
+    return replace(s, org, replace)
+}
+
+fun replace(x: String, replaceChar: CharArray, replaceWith: Array<String>): String {
+    // common case no replacement
+    var ok = true
+    for (aReplaceChar in replaceChar) {
+        val pos = x.indexOf(aReplaceChar)
+        ok = pos < 0
+        if (!ok) break
+    }
+    if (ok) return x
+
+    // gotta do it
+    val sb = StringBuilder(x)
+    for (i in replaceChar.indices) {
+        val pos = x.indexOf(replaceChar[i])
+        if (pos >= 0) {
+            replace(sb, replaceChar[i], replaceWith[i])
+        }
+    }
+    return sb.toString()
+}
+
+fun replace(sb: StringBuilder, remove: Char, replaceWith: String) {
+    var i = 0
+    while (i < sb.length) {
+        if (sb[i] == remove) {
+            sb.replace(i, i + 1, replaceWith)
+            i += replaceWith.length - 1
+        }
+        i++
+    }
+}
